SCD: Start with multiple seed nodes
This adds two new methods to the interface of SelectiveCommunityDetector. The first is for expanding a community with a single seed, which was implemented by all existing algorithms anyway. The second expands a single community starting with multiple seed nodes. Further, the base class gets implementations of all but the second new method so child classes need to implement only one method.

Direct use cases for this method will be implemented in other commits. The general idea is that by specifying more than one seed node one can more precisely specify which community should be expanded and direct the algorithm towards a specific community the node is part of in the case of overlapping communities. In an interactive scenario, a user might also simply select multiple nodes that should be part of the community.